Add redirect and reorganize file structure
