--- client/src/App.graphql
@@ -1,3 +1,0 @@
-query getHistograms {
-  listHistograms
-}
