+++ client/src/components/EMS.graphql
@@ -1,0 +1,3 @@
+query getHistograms {
+  listHistograms
+}
