+++ client/src/index.tsx
@@ -1,17 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
 import { Provider, Client, defaultExchanges } from "urql";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { BrowserRouter, Route, Routes, Navigate } from "react-router-dom";
 
 import {
   EMSPage,
   RunSchedulerPage,
   ControlPanelPage,
   LoginPage,
-} from "./routes";
+} from "./pages";
 import Navigation from "./components/Navigation";
 
 // If you want to start measuring performance in your app, pass a function
@@ -33,7 +32,7 @@ root.render(
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Navigation />}>
-            <Route path="/" element={<App />} />
+            <Route path="/" element={<Navigate replace to="/ems" />} />
             <Route path="ems" element={<EMSPage />} />
             <Route path="run-scheduler" element={<RunSchedulerPage />} />
             <Route path="control-panel" element={<ControlPanelPage />} />
