+++ client/src/pages/EMSPage.tsx
@@ -1,0 +1,13 @@
+//TODO: Set up path shortcuts
+import EMS from "../components/EMS";
+
+const EMSPage = () => {
+  return (
+    <div>
+      This is the EMS Page
+      <EMS />
+    </div>
+  );
+};
+
+export default EMSPage;
