--- client/src/routes/EMSPage.tsx
@@ -1,5 +1,0 @@
-const EMSPage = () => {
-  return <div>This is the EMS Page</div>;
-};
-
-export default EMSPage;
